+++ source/beam_calculation/envelope_3d/__init__.py
@@ -1,5 +1,4 @@
-"""
-This subpackage holds the modules for the :class:`.Envelope3D` beam calculator.
+"""Define modules for the :class:`.Envelope3D` beam calculator.
 
 It is a 3d envelope calculator, only implemented in Python for now.