+++ source/beam_calculation/tracewin/transfer_matrix_factory.py
@@ -1,25 +1,25 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """Provide an easy way to generate :class:`.TransferMatrix`."""
-from typing import Callable
+
 import logging
 import os
+from typing import Callable
 
 import numpy as np
 
-from core.transfer_matrix.transfer_matrix import TransferMatrix
 from core.transfer_matrix.factory import TransferMatrixFactory
+from core.transfer_matrix.transfer_matrix import TransferMatrix
 from tracewin_utils import load
 
 
 class TransferMatrixFactoryTraceWin(TransferMatrixFactory):
     """Provide a method for easy creation of :class:`.TransferMatrix`."""
 
-    def _load_transfer_matrices(self,
-                                path_cal: str,
-                                filename: str = 'Transfer_matrix1.dat',
-                                high_def: bool = False
-                                ) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
+    def _load_transfer_matrices(
+        self,
+        path_cal: str,
+        filename: str = "Transfer_matrix1.dat",
+        high_def: bool = False,
+    ) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
         """
         Get the full transfer matrices calculated by TraceWin.
 
@@ -44,21 +44,25 @@ class TransferMatrixFactoryTraceWin(TransferMatrixFactory):
 
         """
         if high_def:
-            logging.error("High definition not implemented. Can only import"
-                          "transfer matrices @ element positions.")
+            logging.error(
+                "High definition not implemented. Can only import transfer "
+                "matrices @ element positions."
+            )
             high_def = False
 
         path = os.path.join(path_cal, filename)
-        elements_numbers, position_in_m, transfer_matrices = \
+        elements_numbers, position_in_m, transfer_matrices = (
             load.transfer_matrices(path)
+        )
         logging.debug(f"successfully loaded {path}")
         return elements_numbers, position_in_m, transfer_matrices
 
-    def run(self,
-            tm_cumul_in: np.ndarray,
-            path_cal: str,
-            element_to_index: Callable,
-            ) -> TransferMatrix:
+    def run(
+        self,
+        tm_cumul_in: np.ndarray,
+        path_cal: str,
+        element_to_index: Callable,
+    ) -> TransferMatrix:
         r"""Load the TraceWin transfer matrix file and create the object.
 
         Parameters
@@ -82,5 +86,6 @@ class TransferMatrixFactoryTraceWin(TransferMatrixFactory):
             self.is_3d,
             first_cumulated_transfer_matrix=tm_cumul_in,
             cumulated=cumulated,
-            element_to_index=element_to_index)
+            element_to_index=element_to_index,
+        )
         return transfer_matrix