+++ source/config/beam_calculator.py
@@ -1,11 +1,10 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """Define the functions to test the :class:`.BeamCalculator` arguments.
 
 .. todo::
     Handling of arguments for TW could be cleaner
 
 """
+
 import logging
 import socket
 import tomllib
@@ -14,8 +13,8 @@ from pathlib import Path
 from config.helper import check_type
 
 IMPLEMENTED_BEAM_CALCULATORS = ("Envelope1D", "TraceWin", "Envelope3D")  #:
-IMPLEMENTED_ENVELOPE1D_METHODS = ("leapfrog", "RK")  #:
-IMPLEMENTED_ENVELOPE3D_METHODS = ("RK",)  #:
+IMPLEMENTED_ENVELOPE1D_METHODS = ("leapfrog", "RK4")  #:
+IMPLEMENTED_ENVELOPE3D_METHODS = ("RK4",)  #:
 
 IMPLEMENTED_TRACEWIN_ARGUMENTS = (
     "hide",
@@ -119,6 +118,12 @@ def _test_envelope1d(
     **beam_calculator_kw: bool | str | int,
 ) -> None:
     """Test the consistency for the basic :class:`.Envelope1D` beam calculator."""
+    if method == "RK":
+        logging.warning(
+            f"{method = } is deprecated. Prefer 'RK4' (same thing but more "
+            "explicit)."
+        )
+        method = "RK4"
     assert method in IMPLEMENTED_ENVELOPE1D_METHODS
     check_type(bool, "beam_calculator", flag_cython, flag_phi_abs)
     check_type(int, "beam_calculator", n_steps_per_cell)
@@ -128,7 +133,8 @@ def _edit_configuration_dict_in_place_envelope1d(
     beam_calculator_kw: dict, config_folder: Path
 ) -> None:
     """Modify the kw dict inplace."""
-    pass
+    if beam_calculator_kw["method"] == "RK":
+        beam_calculator_kw["method"] = "RK4"
 
 
 def _test_tracewin(
@@ -279,11 +285,17 @@ def _find_file(config_folder: Path, file: str) -> Path:
 def _test_envelope3d(
     flag_phi_abs: bool,
     n_steps_per_cell: int,
-    method: str = "RK",
+    method: str = "RK4",
     flag_cython: bool | None = None,
     **beam_calculator_kw: bool | str | int,
 ) -> None:
     """Check validity of arguments for :class:`.Envelope3D`."""
+    if method == "RK":
+        logging.warning(
+            f"{method = } is deprecated. Prefer 'RK4' (same thing but more "
+            "explicit)."
+        )
+        method = "RK4"
     assert method in IMPLEMENTED_ENVELOPE3D_METHODS
     check_type(bool, "beam_calculator", flag_phi_abs)
     check_type(int, "beam_calculator", n_steps_per_cell)
@@ -296,4 +308,5 @@ def _edit_configuration_dict_in_place_envelope3d(
     beam_calculator_kw: dict, config_folder: Path
 ) -> None:
     """Modify the kw dict inplace."""
-    pass
+    if beam_calculator_kw.get("method", None) == "RK":
+        beam_calculator_kw["method"] = "RK4"