+++ tests/test_beam_calculator/test_envelope1d.py
@@ -29,9 +29,9 @@ leapfrog_marker = pytest.mark.xfail(
 )
 
 params = [
-    pytest.param(("RK", False, False, 40), marks=pytest.mark.smoke),
-    pytest.param(("RK", False, True, 40), marks=pytest.mark.smoke),
-    pytest.param(("RK", True, False, 40), marks=pytest.mark.cython),
+    pytest.param(("RK4", False, False, 40), marks=pytest.mark.smoke),
+    pytest.param(("RK4", False, True, 40), marks=pytest.mark.smoke),
+    pytest.param(("RK4", True, False, 40), marks=pytest.mark.cython),
     pytest.param(("leapfrog", False, False, 60), marks=leapfrog_marker),
     pytest.param(
         ("leapfrog", True, False, 60),